The control of RGT app not working with Wahoo Direct Connect for RGT setting #1055

+++ src/characteristicwriteprocessore005.h
@@ -20,7 +20,7 @@ class CharacteristicWriteProcessorE005 : public CharacteristicWriteProcessor {
     void changePower(uint16_t power);
 
   private:
-    uint16_t weight, rrc, wrc;
+    double weight, rrc, wrc;
 
   signals:
     void changeInclination(double grade, double percentage);
